Pictures and Updates
Added the long awaited example pictures for two steps.

Updated readme to resolve issue #2. Also updated it to include more information and fix typos.

Remodeled the minimum layer speed cone to prevent missing layer errors when slicing and so that the cone actually comes out round and not 20-something sided.